+++ frontend/src/components/DesktopPage/Desktop/scratchPad.js
@@ -1,6 +1,6 @@
 import { useEffect, useState, useRef } from "react";
 import { useSelector, useDispatch } from "react-redux";
-import * as sessionActions from "../../store/session";
+import * as sessionActions from "../../../store/session";
 
 const ScratchPad = () => {
 	const dispatch = useDispatch();
