+++ frontend/src/components/DesktopPage/Trash/index.js
